merge acceptor and server; diverge connection to each worker
This is a big change. 1, merge acceptor and server, 2, do not put all connections in acceptor, because acceptor is in main thread but connections are all in worker thread. So put related connections into each worker. 3, add wakeup event fd to wake up IO thread, or the thread just wait for epoll_wait timeout.